requirements.txt
# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Attachment Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
Pillow>=10.0.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Utilities
requests>=2.31.0
pathlib>=1.0.1

# Development
pytest>=7.4.0
pytest-playwright>=0.4.0
